Try to fix int

+++ lab_11_01_01/inc/specifier_functions.h
@@ -5,7 +5,7 @@
 
 void write_symbol(char *dst, char symbol, size_t max_length, int *cur_length);
 void string_specifier(char *dst, char *src, size_t max_length, int *cur_length);
-int digit_to_string(char *char_digit, int max_len, int digit);
+int digit_to_string(char *char_digit, int max_len, long int digit);
 void digit_specifier(char *dst, int digit, size_t max_length, int *cur_length);
 void hex_specifier(char *dst, unsigned int digit, size_t max_length, int *cur_length);
 int digit_to_hex_string(char *char_digit, int max_len, unsigned int digit);
